fix-dhcp-hack: run service late and restart ssh
This script is a hail mary to force-enable DHCP on your
interfaces should anything have gone south. As such we
cannot want to run this early, as we'd be trying to fix
the race which networking folks care about. We just want
our network interface up. To do this just wait until we
are sure all interfaces are exposed, then update the
interfaces, restart networking and then ssh.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>